requirements.txt
# Core ML libraries
torch>=2.0.0
transformers>=4.30.0
datasets>=2.14.0

# OpenAI API (for GPT-4.1 nano)
openai>=1.0.0

# Configuration and utilities
pyyaml>=6.0
python-dotenv>=1.0.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# Evaluation and metrics
scikit-learn>=1.3.0

# Jupyter for notebooks
jupyter>=1.0.0
ipykernel>=6.25.0

# Logging and monitoring
tqdm>=4.65.0
wandb>=0.15.0  # Optional: for experiment tracking
